Chore: Navigation Component Refactor
